refactor: Cycle5 Step2 - クリーンアーキテクチャ改善
- useScheduleManagementをuseSchedulesに統合（CRUD一元化）
- useAdherenceTrendsのrefetch型をPromise<void>に修正
- 曜日ラベル定数をlib/constants.tsに集約（DRY原則）
- AppointmentFormからMemberEntity直接生成を除去（レイヤー分離）
- useMedicationSearchにerror状態を追加
