+++ src/lib/constants.ts
@@ -1,0 +1,6 @@
+/**
+ * 共通定数
+ */
+
+/** 曜日ラベル（日曜始まり、Date.getDay()のインデックスに対応） */
+export const DAY_LABELS_JP = ['日', '月', '火', '水', '木', '金', '土'] as const;
